Adding Karma Tests to ProjectHub
We have neglected adding Karma Tests to our project -- this means we can
easily edit code somewhere and our edits may affect other parts of the
site.

Actions taken:
* Installed Karma  (via npm)
* Installed Jasmine (via npm)
* Installed angular-mocks (via bower, because it needs to be in sync
with angular which is installed via bower).
* Added a Test folder with an example test.

NOTE: Protractor is not added in this commit. That may come later in
this branch or another.